requirements.txt
Flask==2.3.3 
Flask-SQLAlchemy==3.1.1 
Flask-Migrate==4.0.5 
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
PyMySQL 
gunicorn 
click 
Werkzeug 

pytest
pytest-flask